feat: :construction: refatora e começa a mockar as funcionalidades
